공지사항 생성

+++ routes/web.php
@@ -13,8 +13,8 @@ Route::get('/', function () {
 
 Route::prefix('/admin/product')->group(function(){
     Route::inertia('/create', 'Admin/Product/Index')->name('admin.product.create');
-    Route::get('/manage', [ProductController::class, 'adminIndex'])->name('admin.product.manage'); //
-    Route::post('/create', [ProductController::class, 'store'])->name('admin.product.store'); //
+    Route::get('/manage', [ProductController::class, 'adminIndex'])->name('admin.product.manage'); 
+    Route::post('/create', [ProductController::class, 'store'])->name('admin.product.store'); 
 });
 
 Route::prefix('/user/k-coin')->group(function(){
@@ -26,9 +26,11 @@ Route::prefix('/admin/k-coin')->group(function(){
     Route::get('/manage', [CoinController::class, 'adminManageIndex'])->name('admin.coin.manage'); 
 });
 
-Route::get('/inquiry', [InquiryController::class, 'index'])->name('inquiry');
-
-Route::prefix('/announcement')->group(function(){
-    Route::get('/', [AnnouncementController::class, 'index'])->name('announcement');
-    Route::delete('/delete/{id}', [AnnouncementController::class, 'destroy'])->name('announcement.delete');
+Route::prefix('/admin/announcement')->group(function(){
+    Route::get('/create', [AnnouncementController::class, 'create'])->name('admin.announcement.create');
+    Route::delete('/delete/{id}', [AnnouncementController::class, 'destroy'])->name('admin.announcement.delete');
+    Route::post('/create', [AnnouncementController::class, 'store'])->name('admin.announcement.store');
 });
+
+Route::get('/inquiry', [InquiryController::class, 'index'])->name('inquiry');
+Route::get('/announcement', [AnnouncementController::class, 'index'])->name('announcement');
